+++ cSDDS/Memory.h
@@ -25,7 +25,7 @@ typedef uint8_t BYTE;
 /// <summary>
 /// A smart strlen function that will return 0 for a null pointer
 /// </summary>
-size_t cStrLen(char *s);
+uint32_t cStrLen(char *s);
 
 /// <summary>
 /// Appends newStr to origStr. Returns true on success.
